regulatory-ai-feature-spark-job/src/main/requirements.txt
Cython
docutils
entrypoints
Faker
iniconfig
ipykernel
ipython
kafka-python
MarkupSafe
numpy
oauthlib
pandas
py
pytest
python-apt
python-edgar
PyYAML
regex
requests
scipy
urllib3